Changes to the OCELOT repository
- xfel_utils (OCELOT/utils/xfel_utils.py): add the case when there is no
lattice file created (using the lattice parameters within the input
file)
- genesis.py (OCELOT/adaptors/genesis.py):  Remove the write_input_file
function that does not have any use
- genesis_plot(OCELOT/gui/genesis_plot.py)  Add modifications to
gen_outplot_statistics so that the outcome of the GENESIS simulations
for several noise realisations can be plotted. Modifications to the
sub_subfig_und_quad, subfig_el_size, subfig_rad_pow, so that the user
can have control on the legend and colour of the lines. for the
subfig_rad_pow function, a flag was introduced in case the user wants to
plot in logarithmic scale. For the gen_outplot_dfl function, a new input
parameter was added (filename), in case the user wants to rename the dfl
obtained plots with such filename